Added otp verification, Payment gateway, Added Loader(Not working right now), fixed login popup bug

+++ backend/controllers/orderController.js
@@ -1,11 +1,10 @@
 import orderModel from "../models/orderModel.js";
 import userModel from "../models/userModel.js";
-// import Stripe from "stripe"
-
-// const stripe = new stripe(process.env.STRIPE_SECRET_KEY)
+import Razorpay from "razorpay";
+import crypto from "crypto";
 
 const placeOrder = async (req, res) => {
-    const frontend_url = "https://food-delivery-website-frontend-ndjx.onrender.com";
+    const frontend_url = "http://localhost:5174";
 
     try {
         const newOrder = new orderModel({
@@ -18,40 +17,65 @@ const placeOrder = async (req, res) => {
         await newOrder.save();
         await userModel.findByIdAndUpdate(req.body.userId, { cartData: {} });
 
-        // 7:56:06
         const success_url = `${frontend_url}/verify?success=true&orderId=${newOrder._id}`;
 
-        res.json({success: true, session_url: success_url})
+        const instance = new Razorpay({
+            key_id: process.env.RAZORPAY_KEY_ID,
+            key_secret: process.env.RAZORPAY_KEY_SECRET,
+        })
+
+        const options = {
+            amount: req.body.amount * 100,
+            currency: "INR",
+            receipt: crypto.randomBytes(10).toString("hex"),
+        }
+
+        instance.orders.create(options, (error, order) => {
+            if (error) {
+                console.log(error);
+                res.status(500).json({ success: false, payment: "failed" })
+            }
+            res.status(200).json({ success: true, data: order });
+            // res.json({ success: true, session_url: success_url })
+        })
     } catch (error) {
         console.log(error);
-        res.json({success: false, message: "Error"})
+        res.status(500).json({ success: false, message: "Internal Server Error !!!" });
     }
 }
 
 const verifyOrder = async (req, res) => {
-    const { orderId } = req.body;
     try {
-        if(true) {
+        const { razorpay_order_id, razorpay_payment_id, razorpay_signature } = req.body;
+        const sign = razorpay_order_id + "|" + razorpay_payment_id;
+
+        const expectedSign = crypto
+            .createHmac("sha256", process.env.RAZORPAY_KEY_SECRET)
+            .update(sign.toString())
+            .digest("hex");
+
+        if (expectedSign == razorpay_signature) {
             await orderModel.findByIdAndUpdate(orderId, { payment: true });
-            res.json({success: true, message: "Paid"});
-        } /*else {
+            return res.status(200).json({ success: true, message: "Payment Successfully!" });
+        } else {
             await orderModel.findByIdAndDelete(orderId);
-            res.json({success: false, message: "Payment Failed"});
-        }*/
+            res.json({ success: false, message: "Payment Failed" });
+        }
+
     } catch (error) {
         console.log(error);
-        res.json({success: false, message: "Error"});
+        res.json({ success: false, message: "Error" });
     }
 }
 
 // user orders for frontend
 const userOrders = async (req, res) => {
     try {
-        const orders = await orderModel.find({userId: req.body.userId})
-        res.json({success: true, data: orders});
+        const orders = await orderModel.find({ userId: req.body.userId })
+        res.json({ success: true, data: orders });
     } catch (error) {
         console.log(error);
-        res.json({success: false, message: "Error"})
+        res.json({ success: false, message: "Error" })
     }
 }
 
@@ -59,21 +83,21 @@ const userOrders = async (req, res) => {
 const listOrders = async (req, res) => {
     try {
         const orders = await orderModel.find({});
-        res.json({success: true, data: orders})
+        res.json({ success: true, data: orders })
     } catch (error) {
         console.log(error);
-        res.json({success: false, message: "Error"})        
+        res.json({ success: false, message: "Error" })
     }
 }
 
 // api for updating order status
 const updateStatus = async (req, res) => {
     try {
-        await orderModel.findByIdAndUpdate(req.body.orderId, {status: req.body.status})
-        res.json({success: true, message: "Status updated"})
+        await orderModel.findByIdAndUpdate(req.body.orderId, { status: req.body.status })
+        res.json({ success: true, message: "Status updated" })
     } catch (error) {
         console.log(error);
-        res.json({success: false, message: "Error"})
+        res.json({ success: false, message: "Error" })
     }
 }
